fix secOrder

+++ src/provider-goods/model/goods.go
@@ -18,7 +18,7 @@ type Goods struct {
 	Price decimal.Decimal
 
 	//stock 商品库存
-	Stock int64
+	Stock int
 
 	//photo 商品图片
 	Photo string `gorm:"type:varchar(255);not null"`
